no error for competition fragment
