+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,6 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".view.activities.CompetitionActivity" />
         <activity android:name=".view.activities.TeamActivity" />
         <activity android:name=".view.activities.PlayerActivity" />
         <activity android:name=".view.activities.CreditsActivity" />
