+++ app/src/main/java/com/example/footballapi/model/model_dao/AdapterRV_Search.java
@@ -13,8 +13,8 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.example.footballapi.R;
-import com.example.footballapi.view.activities.CompetitionActivity;
 import com.example.footballapi.view.activities.TeamActivity;
+import com.example.footballapi.view.fragments.CompetitionFragment;
 
 import java.util.List;
 
@@ -76,7 +76,7 @@ public class AdapterRV_Search extends RecyclerView.Adapter<AdapterRV_Search.View
             @Override
             public void onClick(View v) {
                 Context context = v.getContext();
-                Intent intent = new Intent(context, CompetitionActivity.class);
+                Intent intent = new Intent(context, CompetitionFragment.class);
                 intent.putExtra(CLE_DONNEES_ID_COMPET, listSearch.get(position).getIdCompet());
                 context.startActivity(intent);
             }
