+++ app/src/main/java/com/example/footballapi/view/activities/MainActivity.java
@@ -12,6 +12,7 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 
 import com.example.footballapi.R;
+import com.example.footballapi.view.fragments.CompetitionFragment;
 import com.google.android.material.navigation.NavigationView;
 
 public class MainActivity extends AppCompatActivity {
