+++ app/src/main/java/com/example/footballapi/view/fragments/ClassementFragment.java
@@ -13,7 +13,6 @@ import com.example.footballapi.R;
 import com.example.footballapi.controleur.ClassementController;
 import com.example.footballapi.model.model_recyclerview.classement.AdapterRV_Classement;
 import com.example.footballapi.model.model_recyclerview.classement.TeamModel;
-import com.example.footballapi.view.activities.CompetitionActivity;
 
 import java.util.HashMap;
 import java.util.List;
@@ -50,7 +49,7 @@ public class ClassementFragment extends Fragment {
     }
 
     public void showList(List<TeamModel> list, boolean netaccess, HashMap<Integer, String> teamNameCrest){
-        if (teamNameCrest != null) CompetitionActivity.setTeamsNameCrests(teamNameCrest);
+        if (teamNameCrest != null) CompetitionFragment.setTeamsNameCrests(teamNameCrest);
 
         layoutManager = new LinearLayoutManager(getContext());
         rvClassement.setLayoutManager(layoutManager);
