+++ app/src/main/res/layout/fragment_competition.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.activities.CompetitionActivity"
+    tools:context=".view.fragments.CompetitionFragment"
     android:orientation="vertical">
 
     <androidx.viewpager.widget.ViewPager
